services/tiler-api/requirements.txt
titiler.pgstac==2.1.0
fastapi>=0.119.0
uvicorn>=0.30.1
psycopg[binary]
psycopg-pool
